Some refactoring for custom method hooks
