+++ package.json
@@ -50,7 +50,7 @@
     "node": ">= 6"
   },
   "dependencies": {
-    "@feathersjs/commons": "^2.0.0",
+    "@feathersjs/commons": "^3.0.0-pre.1",
     "debug": "^3.1.0",
     "events": "^3.0.0",
     "uberproto": "^2.0.2"
